Feat: Outsourced content to config  file (unfinished)
