+++ components/Hero/Hero.js
@@ -1,16 +1,18 @@
 import styles from "./Hero.module.css";
 
+import { title, subtitle } from "../../config";
+
 import { Text } from "@nextui-org/react";
 
 export default function Hero() {
   return (
     <div className={styles["main-title__container"]}>
       <Text h1 size={60} weight="bold">
-        Hi, I'm Javi
+        {title}
       </Text>
       <div className={styles["main-title__subtitle"]}>
         <Text h3 size={25} weight="light">
-          Engineer, Developer and Curious
+          {subtitle}
         </Text>
       </div>
     </div>
