+++ components/Navbar/Navbar.js
@@ -1,23 +1,24 @@
 import styles from "./Navbar.module.css";
 
+import { title } from "../../config";
+
 import { AiFillGithub } from "react-icons/ai";
 import {
   HiOutlineDocumentText,
   HiOutlineArrowNarrowRight,
 } from "react-icons/hi";
-import { BsBook } from "react-icons/bs";
 import { FaLinkedin, FaTwitter } from "react-icons/fa";
 
 import { Button, Text } from "@nextui-org/react";
 
-import { externalLinks } from "../../constants";
+import { externalLinks } from "../../config";
 
 export default function Navbar() {
   return (
     <nav className={styles["navbar-container"]}>
       <div className={styles["navbar-brand"]}>
         <Text h1 size={24} weight="bold">
-          Hi, I'm Javi
+          {title}
         </Text>
       </div>
       <div className={styles["navbar-links__container"]}>
@@ -38,28 +39,28 @@ export default function Navbar() {
           </li>
           <li className={styles["navbar-link__list-item"]}>
             <a href={externalLinks.github} className={styles["navbar-link"]}>
-              <Button flat color="primary" auto>
+              <Button light color="primary" auto>
                 <AiFillGithub /> &nbsp;Github
               </Button>
             </a>
           </li>
           <li className={styles["navbar-link__list-item"]}>
             <a href={externalLinks.resume} className={styles["navbar-link"]}>
-              <Button flat color="primary" auto>
+              <Button light color="primary" auto>
                 <HiOutlineDocumentText /> &nbsp;Resume
               </Button>
             </a>
           </li>
           <li className={styles["navbar-link__list-item"]}>
             <a href={externalLinks.linkedin} className={styles["navbar-link"]}>
-              <Button flat color="primary" auto>
+              <Button light color="primary" auto>
                 <FaLinkedin /> &nbsp;LinkedIn
               </Button>
             </a>
           </li>
           <li className={styles["navbar-link__list-item"]}>
             <a href={externalLinks.twitter} className={styles["navbar-link"]}>
-              <Button flat color="primary" auto>
+              <Button light color="primary" auto>
                 <FaTwitter /> &nbsp;Twitter
               </Button>
             </a>
@@ -69,7 +70,7 @@ export default function Navbar() {
               href="https://blog.javisanchez.me/"
               className={styles["navbar-link"]}
             >
-              <Button flat color="primary" auto>
+              <Button light color="primary" auto>
                 Blog&nbsp; <HiOutlineArrowNarrowRight />
               </Button>
             </a>
