--- constants.js
@@ -1,8 +1,0 @@
-export const externalLinks = {
-  github: "https://github.com/sanchezfdezjavier",
-  resume:
-    "https://docs.google.com/document/d/1yjZM0yORhAmvz0hcGOjrZ4MK1KgP68buQTprcvOTkzM/edit?usp=sharing",
-  linkedin: "https://www.linkedin.com/in/sanchezfdezjavier/",
-  blog: "https://blog.javisanchez.me/",
-  twitter: "https://twitter.com/javisnchztech",
-};
